+++ control-plane/agpctl/internal/options/options.go
@@ -4,13 +4,30 @@
 package options
 
 import (
+	"time"
+
 	"go.uber.org/zap"
 )
 
 // CommonOptions provides the common flags, options, for all the commands.
 type CommonOptions struct {
-	// The configuration file, if used
-	ConfigFile string
+	// AGP Gateway control API endpoint
+	Server string
+
+	// gRPC request timeout
+	Timeout time.Duration
+
+	// TLS certificate verification
+	TLSInsecure bool
+
+	// CA certificate
+	TLSCAFile string
+
+	// client TLS certificate
+	TLSCertFile string
+
+	// client TLS key
+	TLSKeyFile string
 
 	// The logger. For now we only use zap
 	Logger *zap.Logger