+++ data-plane/gateway/datapath/build.rs
@@ -13,9 +13,6 @@ fn main() {
 
     tonic_build::configure()
         .out_dir("src/pubsub/gen")
-        .compile_protos(
-            &["../../../proto/pubsub/v1/pubsub.proto"],
-            &["../../../proto"],
-        )
+        .compile_protos(&["proto/v1/pubsub.proto"], &["proto/v1"])
         .unwrap();
 }