+++ data-plane/gateway/datapath/src/messages/encoder.rs
@@ -7,12 +7,34 @@ use crate::pubsub::ProtoAgent;
 
 pub const DEFAULT_AGENT_ID: u64 = u64::MAX;
 
-#[derive(Hash, Eq, PartialEq, Debug, Clone, Default)]
+#[derive(Debug, Clone, Default)]
 pub struct AgentType {
     organization: u64,
     namespace: u64,
     agent_type: u64,
+
+    // Store the original string representation of the agent type
+    // This is useful for debugging and logging purposes
+    strings: Option<Box<(String, String, String)>>,
 }
+
+impl Hash for AgentType {
+    fn hash<H: Hasher>(&self, state: &mut H) {
+        self.organization.hash(state);
+        self.namespace.hash(state);
+        self.agent_type.hash(state);
+    }
+}
+
+impl PartialEq for AgentType {
+    fn eq(&self, other: &Self) -> bool {
+        self.organization == other.organization
+            && self.namespace == other.namespace
+            && self.agent_type == other.agent_type
+    }
+}
+
+impl Eq for AgentType {}
 
 impl std::fmt::Display for AgentType {
     fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
@@ -20,7 +42,13 @@ impl std::fmt::Display for AgentType {
             f,
             "{:x}/{:x}/{:x}",
             self.organization, self.namespace, self.agent_type
-        )
+        )?;
+
+        if let Some(strings) = &self.strings {
+            write!(f, " ({}/{}/{})", strings.0, strings.1, strings.2)?;
+        }
+
+        Ok(())
     }
 }
 
@@ -30,6 +58,7 @@ impl From<&ProtoAgent> for AgentType {
             organization: agent.organization,
             namespace: agent.namespace,
             agent_type: agent.agent_type,
+            strings: None,
         }
     }
 }
@@ -41,6 +70,7 @@ impl AgentType {
             organization,
             namespace,
             agent_type,
+            strings: None,
         }
     }
 
@@ -49,6 +79,11 @@ impl AgentType {
             organization: calculate_hash(organization),
             namespace: calculate_hash(namespace),
             agent_type: calculate_hash(agent_type),
+            strings: Some(Box::new((
+                organization.to_string(),
+                namespace.to_string(),
+                agent_type.to_string(),
+            ))),
         }
     }